src/catflap/requirements_pi.txt
tflite-runtime==2.11.0
tflite-support==0.4.3
opencv-python==4.7.0.68
numpy==1.23.4
matplotlib
python-statemachine==2.0.0
# Raspberry Pi Only
#RPi.GPIO
#picamera2

# sudo apt install -y python3-pyqt5 python3-opengl
# sudo apt install -y python3-picamera2
# sudo apt install libcap-dev
# sudo apt install -y python3-libcamera
# cp -R /usr/lib/python3/dist-packages/libcamera/ ./venv/lib/python3.9/site-packages/
# cp -R /usr/lib/python3/dist-packages/pykms/ ./venv/lib/python3.9/site-packages/

